+++ Assets/Scripts/Shared/IDamageable.cs
@@ -6,6 +6,6 @@ public interface IDamagable
     public TeamID getTeamID();
     public bool getIsAlive();
     public bool getIsBase();
-    public void TakeDamage(int dmg){}
+    public void TakeDamage(float dmg){}
     private void DeathHandler(){}
 }
